Use createElement(s) helper functions.
Use two helpers: createElement (for a single HTML node) and
createElements (for a HTML node which contains other nodes).  The
names are similar but short, and although they can take the same
amount of arguments (each has a 2 argument flavor), the context
should make it clear what is trying to be accomplished.

The similar names are also less unwieldy than qualifying one or
the other with something like "Empty".

Code size and complexity is reduced quite a bit with these helpers,
and the node order is more readily apparently when using the list
flavor of createElements().