Make checkbox selections match across all views.
DOM ids are unique, so selecting a checkbox in the menu or the search
would not necessarily mean the other checkbox would match.  Use names
instead (which can be shared by multiple elements) so that all
package selection state checkboxes can be made to match whenever one
of their states changes.